added data Encoding library which had been missed changed sampler id size to 16bit due to inability to pack smaller parser reads packets naively

+++ code/Small_scale/Sampler.X/Sampler.h
@@ -31,7 +31,7 @@ typedef struct XYZPoint_t {
 typedef union MagAccelSet_t {
 
     struct {
-        uint8_t PacketId;
+        uint16_t PacketId;
         XYZPoint_t AccelData[10];
         XYZPoint_t MagData;
     } DataAccess;
